More file restructuring
