+++ test/basic_rng/benchmark_speed.cc
