+++ test/basic_rng/test_continuous.cc
