+++ test/basic_rng/test_discrete.cc
