+++ test/basic_rng/test_distortion.cc
