+++ test/linops/test_sketch_symmetric.cc
