+++ test/linops/test_sketch_vector.cc
