+++ test/linops/test_sparse_trsm.cc
