+++ test/linops/test_spmm/spmm_test_helpers.hh
@@ -30,7 +30,7 @@
 #pragma once
 
 #include "RandBLAS/testing/sparse_data_common.hh"
-#include "test/test_matmul_cores/linop_common.hh"
+#include "test/linops/linop_common.hh"
 #include <gtest/gtest.h>
 
 using namespace test::linop_common;
