+++ test/linops/test_spmm/test_spmm_csc.cc
@@ -1,4 +1,4 @@
-// Copyright, 2026. See LICENSE for copyright holder information.
+// Copyright, 2024. See LICENSE for copyright holder information.
 //
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are met:
@@ -27,34 +27,33 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-#include "test/test_matmul_cores/test_spmm/spmm_test_helpers.hh"
+#include "test/linops/test_spmm/spmm_test_helpers.hh"
 #include <algorithm>
 #include <vector>
-
 using RandBLAS::testing::iid_sparsify_random_dense;
 
 using namespace RandBLAS::sparse_data;
-using namespace RandBLAS::sparse_data::csr;
+using namespace RandBLAS::sparse_data::csc;
 using blas::Layout;
 
 
 template <typename T>
-class TestLeftMultiply_CSR : public TestLeftMultiply_Sparse<CSRMatrix<T>> {
-    CSRMatrix<T> make_test_matrix(int64_t m, int64_t n, T nonzero_prob, uint32_t key = 0) {
+class TestLeftMultiply_CSC : public TestLeftMultiply_Sparse<CSCMatrix<T>> {
+    CSCMatrix<T> make_test_matrix(int64_t m, int64_t n, T nonzero_prob, uint32_t key = 0) {
         randblas_require(nonzero_prob >= 0);
         randblas_require(nonzero_prob <= 1);
-        CSRMatrix<T> A(m, n);
+        CSCMatrix<T> A(m, n);
         std::vector<T> actual(m * n);
         RandBLAS::RNGState s(key);
         iid_sparsify_random_dense<T>(m, n, Layout::ColMajor, actual.data(), 1 - nonzero_prob, s);
-        dense_to_csr<T>(Layout::ColMajor, actual.data(), 0.0, A);
+        dense_to_csc<T>(Layout::ColMajor, actual.data(), 0.0, A);
         return A;
     }
 };
 
-class TestLeftMultiply_CSR_double : public TestLeftMultiply_CSR<double> {};
+class TestLeftMultiply_CSC_double : public TestLeftMultiply_CSC<double> {};
 
-class TestLeftMultiply_CSR_single : public TestLeftMultiply_CSR<float> {};
+class TestLeftMultiply_CSC_single : public TestLeftMultiply_CSC<float> {};
 
 ////////////////////////////////////////////////////////////////////////
 //
@@ -64,7 +63,7 @@ class TestLeftMultiply_CSR_single : public TestLeftMultiply_CSR<float> {};
 //
 ////////////////////////////////////////////////////////////////////////
 
-TEST_F(TestLeftMultiply_CSR_double, tall_multiply_eye_colmajor) {
+TEST_F(TestLeftMultiply_CSC_double, tall_multiply_eye_colmajor) {
     for (uint32_t key : {0}) {
         multiply_eye(key, 200, 30, Layout::ColMajor, 0.01);
         multiply_eye(key, 200, 30, Layout::ColMajor, 0.10);
@@ -72,7 +71,7 @@ TEST_F(TestLeftMultiply_CSR_double, tall_multiply_eye_colmajor) {
     }
 }
 
-TEST_F(TestLeftMultiply_CSR_double, tall_multiply_eye_rowmajor) {
+TEST_F(TestLeftMultiply_CSC_double, tall_multiply_eye_rowmajor) {
     for (uint32_t key : {0}) {
         multiply_eye(key, 200, 30, Layout::RowMajor, 0.01);
         multiply_eye(key, 200, 30, Layout::RowMajor, 0.10);
@@ -80,7 +79,7 @@ TEST_F(TestLeftMultiply_CSR_double, tall_multiply_eye_rowmajor) {
     }
 }
 
-TEST_F(TestLeftMultiply_CSR_double, wide_multiply_eye_colmajor) {
+TEST_F(TestLeftMultiply_CSC_double, wide_multiply_eye_colmajor) {
     for (uint32_t key : {0}) {
         multiply_eye(key, 51, 101, Layout::ColMajor, 0.01);
         multiply_eye(key, 51, 101, Layout::ColMajor, 0.10);
@@ -88,7 +87,7 @@ TEST_F(TestLeftMultiply_CSR_double, wide_multiply_eye_colmajor) {
     }
 }
 
-TEST_F(TestLeftMultiply_CSR_double, wide_multiply_eye_rowmajor) {
+TEST_F(TestLeftMultiply_CSC_double, wide_multiply_eye_rowmajor) {
     for (uint32_t key : {0}) {
         multiply_eye(key, 51, 101, Layout::RowMajor, 0.01);
         multiply_eye(key, 51, 101, Layout::RowMajor, 0.10);
@@ -96,7 +95,7 @@ TEST_F(TestLeftMultiply_CSR_double, wide_multiply_eye_rowmajor) {
     }
 }
 
-TEST_F(TestLeftMultiply_CSR_double, nontrivial_scales_colmajor1) {
+TEST_F(TestLeftMultiply_CSC_double, nontrivial_scales_colmajor1) {
     double alpha = 5.5;
     double beta = 0.0;
     alpha_beta(0, alpha, beta, 21, 4, Layout::ColMajor, 0.05);
@@ -104,7 +103,7 @@ TEST_F(TestLeftMultiply_CSR_double, nontrivial_scales_colmajor1) {
     alpha_beta(0, alpha, beta, 21, 4, Layout::ColMajor, 0.80);
 }
 
-TEST_F(TestLeftMultiply_CSR_double, nontrivial_scales_colmajor2) {
+TEST_F(TestLeftMultiply_CSC_double, nontrivial_scales_colmajor2) {
     double alpha = 5.5;
     double beta = -1.0;
     alpha_beta(0, alpha, beta, 21, 4, Layout::ColMajor, 0.05);
@@ -112,7 +111,7 @@ TEST_F(TestLeftMultiply_CSR_double, nontrivial_scales_colmajor2) {
     alpha_beta(0, alpha, beta, 21, 4, Layout::ColMajor, 0.80);
 }
 
-TEST_F(TestLeftMultiply_CSR_double, nontrivial_scales_rowmajor1) {
+TEST_F(TestLeftMultiply_CSC_double, nontrivial_scales_rowmajor1) {
     double alpha = 5.5;
     double beta = 0.0;
     alpha_beta(0, alpha, beta, 21, 4, Layout::RowMajor, 0.05);
@@ -120,7 +119,7 @@ TEST_F(TestLeftMultiply_CSR_double, nontrivial_scales_rowmajor1) {
     alpha_beta(0, alpha, beta, 21, 4, Layout::RowMajor, 0.80);
 }
 
-TEST_F(TestLeftMultiply_CSR_double, nontrivial_scales_rowmajor2) {
+TEST_F(TestLeftMultiply_CSC_double, nontrivial_scales_rowmajor2) {
     double alpha = 5.5;
     double beta = -1.0;
     alpha_beta(0, alpha, beta, 21, 4, Layout::RowMajor, 0.05);
@@ -134,7 +133,7 @@ TEST_F(TestLeftMultiply_CSR_double, nontrivial_scales_rowmajor2) {
 //
 ////////////////////////////////////////////////////////////////////////
 
-TEST_F(TestLeftMultiply_CSR_double, transpose_self_colmajor) {
+TEST_F(TestLeftMultiply_CSC_double, transpose_self_colmajor) {
     for (uint32_t key : {0}) {
         transpose_self(key, 200, 30, Layout::ColMajor, 0.01);
         transpose_self(key, 200, 30, Layout::ColMajor, 0.10);
@@ -142,7 +141,7 @@ TEST_F(TestLeftMultiply_CSR_double, transpose_self_colmajor) {
     }
 }
 
-TEST_F(TestLeftMultiply_CSR_double, transpose_self_rowmajor) {
+TEST_F(TestLeftMultiply_CSC_double, transpose_self_rowmajor) {
     for (uint32_t key : {0}) {
         transpose_self(key, 200, 30, Layout::RowMajor, 0.01);
         transpose_self(key, 200, 30, Layout::RowMajor, 0.10);
@@ -150,7 +149,7 @@ TEST_F(TestLeftMultiply_CSR_double, transpose_self_rowmajor) {
     }
 }
 
-TEST_F(TestLeftMultiply_CSR_single, transpose_self) {
+TEST_F(TestLeftMultiply_CSC_single, transpose_self) {
     for (uint32_t key : {0}) {
         transpose_self(key, 200, 30, Layout::ColMajor, 0.01);
         transpose_self(key, 200, 30, Layout::ColMajor, 0.10);
@@ -164,21 +163,21 @@ TEST_F(TestLeftMultiply_CSR_single, transpose_self) {
 //
 ////////////////////////////////////////////////////////////////////////
 
-TEST_F(TestLeftMultiply_CSR_double, submatrix_other_double_colmajor) {
+TEST_F(TestLeftMultiply_CSC_double, submatrix_other_double_colmajor) {
     for (uint32_t key : {0}) {
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 0.1);
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 1.0);
     }
 }
 
-TEST_F(TestLeftMultiply_CSR_double, submatrix_other_double_rowmajor) {
+TEST_F(TestLeftMultiply_CSC_double, submatrix_other_double_rowmajor) {
     for (uint32_t key : {0}) {
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::RowMajor, 0.1);
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::RowMajor, 1.0);
     }
 }
 
-TEST_F(TestLeftMultiply_CSR_double, submatrix_other_single) {
+TEST_F(TestLeftMultiply_CSC_double, submatrix_other_single) {
     for (uint32_t key : {0}) {
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 0.1);
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 1.0);
@@ -192,14 +191,14 @@ TEST_F(TestLeftMultiply_CSR_double, submatrix_other_single) {
 ////////////////////////////////////////////////////////////////////////
 
 
-TEST_F(TestLeftMultiply_CSR_double, sparse_times_trans_other_colmajor) {
+TEST_F(TestLeftMultiply_CSC_double, sparse_times_trans_other_colmajor) {
     uint32_t key = 0;
     transpose_other(key, 7, 22, 5, Layout::ColMajor, 0.05);
     transpose_other(key, 7, 22, 5, Layout::ColMajor, 0.10);
     transpose_other(key, 7, 22, 5, Layout::ColMajor, 0.80);
 }
 
-TEST_F(TestLeftMultiply_CSR_double, sparse_times_trans_other_rowmajor) {
+TEST_F(TestLeftMultiply_CSC_double, sparse_times_trans_other_rowmajor) {
     uint32_t key = 0;
     transpose_other(key, 7, 22, 5, Layout::RowMajor, 0.05);
     transpose_other(key, 7, 22, 5, Layout::RowMajor, 0.10);
@@ -208,23 +207,22 @@ TEST_F(TestLeftMultiply_CSR_double, sparse_times_trans_other_rowmajor) {
 
 
 template <typename T>
-class TestRightMultiply_CSR : public TestRightMultiply_Sparse<CSRMatrix<T>> {
-    CSRMatrix<T> make_test_matrix(int64_t m, int64_t n, T nonzero_prob, uint32_t key = 0) {
+class TestRightMultiply_CSC : public TestRightMultiply_Sparse<CSCMatrix<T>> {
+    CSCMatrix<T> make_test_matrix(int64_t m, int64_t n, T nonzero_prob, uint32_t key = 0) {
         randblas_require(nonzero_prob >= 0);
         randblas_require(nonzero_prob <= 1);
-        CSRMatrix<T> A(m, n);
+        CSCMatrix<T> A(m, n);
         std::vector<T> actual(m * n);
         RandBLAS::RNGState s(key);
         iid_sparsify_random_dense<T>(m, n, Layout::ColMajor, actual.data(), 1 - nonzero_prob, s);
-        dense_to_csr<T>(Layout::ColMajor, actual.data(), 0.0, A);
+        dense_to_csc<T>(Layout::ColMajor, actual.data(), 0.0, A);
         return A;
     }
 };
 
-class TestRightMultiply_CSR_double : public TestRightMultiply_CSR<double> {};
+class TestRightMultiply_CSC_double : public TestRightMultiply_CSC<double> {};
 
-class TestRightMultiply_CSR_single : public TestRightMultiply_CSR<float> {};
-
+class TestRightMultiply_CSC_single : public TestRightMultiply_CSC<float> {};
 
 ////////////////////////////////////////////////////////////////////////
 //
@@ -234,7 +232,7 @@ class TestRightMultiply_CSR_single : public TestRightMultiply_CSR<float> {};
 //
 ////////////////////////////////////////////////////////////////////////
 
-TEST_F(TestRightMultiply_CSR_double, wide_multiply_eye_double_colmajor) {
+TEST_F(TestRightMultiply_CSC_double, wide_multiply_eye_double_colmajor) {
     for (uint32_t key : {0}) {
         multiply_eye(key, 200, 30, Layout::ColMajor, 0.01);
         multiply_eye(key, 200, 30, Layout::ColMajor, 0.10);
@@ -242,7 +240,7 @@ TEST_F(TestRightMultiply_CSR_double, wide_multiply_eye_double_colmajor) {
     }
 }
 
-TEST_F(TestRightMultiply_CSR_double, wide_multiply_eye_double_rowmajor) {
+TEST_F(TestRightMultiply_CSC_double, wide_multiply_eye_double_rowmajor) {
     for (uint32_t key : {0}) {
         multiply_eye(key, 200, 30, Layout::RowMajor, 0.01);
         multiply_eye(key, 200, 30, Layout::RowMajor, 0.10);
@@ -251,7 +249,7 @@ TEST_F(TestRightMultiply_CSR_double, wide_multiply_eye_double_rowmajor) {
 }
 
 
-TEST_F(TestRightMultiply_CSR_double, tall_multiply_eye_double_colmajor) {
+TEST_F(TestRightMultiply_CSC_double, tall_multiply_eye_double_colmajor) {
     for (uint32_t key : {0}) {
         multiply_eye(key, 51, 101, Layout::ColMajor, 0.01);
         multiply_eye(key, 51, 101, Layout::ColMajor, 0.10);
@@ -259,7 +257,7 @@ TEST_F(TestRightMultiply_CSR_double, tall_multiply_eye_double_colmajor) {
     }
 }
 
-TEST_F(TestRightMultiply_CSR_double, tall_multiply_eye_double_rowmajor) {
+TEST_F(TestRightMultiply_CSC_double, tall_multiply_eye_double_rowmajor) {
     for (uint32_t key : {0}) {
         multiply_eye(key, 51, 101, Layout::RowMajor, 0.01);
         multiply_eye(key, 51, 101, Layout::RowMajor, 0.10);
@@ -267,7 +265,7 @@ TEST_F(TestRightMultiply_CSR_double, tall_multiply_eye_double_rowmajor) {
     }
 }
 
-TEST_F(TestRightMultiply_CSR_double, nontrivial_scales_colmajor1) {
+TEST_F(TestRightMultiply_CSC_double, nontrivial_scales_colmajor1) {
     double alpha = 5.5;
     double beta = 0.0;
     alpha_beta(0, alpha, beta, 4, 21, Layout::ColMajor, 0.05);
@@ -275,7 +273,7 @@ TEST_F(TestRightMultiply_CSR_double, nontrivial_scales_colmajor1) {
     alpha_beta(0, alpha, beta, 4, 21, Layout::ColMajor, 0.80);
 }
 
-TEST_F(TestRightMultiply_CSR_double, nontrivial_scales_colmajor2) {
+TEST_F(TestRightMultiply_CSC_double, nontrivial_scales_colmajor2) {
     double alpha = 5.5;
     double beta = -1.0;
     alpha_beta(0, alpha, beta, 4, 21, Layout::ColMajor, 0.05);
@@ -283,7 +281,7 @@ TEST_F(TestRightMultiply_CSR_double, nontrivial_scales_colmajor2) {
     alpha_beta(0, alpha, beta, 4, 21, Layout::ColMajor, 0.80);
 }
 
-TEST_F(TestRightMultiply_CSR_double, nontrivial_scales_rowmajor1) {
+TEST_F(TestRightMultiply_CSC_double, nontrivial_scales_rowmajor1) {
     double alpha = 5.5;
     double beta = 0.0;
     alpha_beta(0, alpha, beta, 4, 21, Layout::RowMajor, 0.05);
@@ -291,7 +289,7 @@ TEST_F(TestRightMultiply_CSR_double, nontrivial_scales_rowmajor1) {
     alpha_beta(0, alpha, beta, 4, 21, Layout::RowMajor, 0.80);
 }
 
-TEST_F(TestRightMultiply_CSR_double, nontrivial_scales_rowmajor2) {
+TEST_F(TestRightMultiply_CSC_double, nontrivial_scales_rowmajor2) {
     double alpha = 5.5;
     double beta = -1.0;
     alpha_beta(0, alpha, beta, 4, 21, Layout::RowMajor, 0.05);
@@ -305,7 +303,7 @@ TEST_F(TestRightMultiply_CSR_double, nontrivial_scales_rowmajor2) {
 //
 ////////////////////////////////////////////////////////////////////////
 
-TEST_F(TestRightMultiply_CSR_double, transpose_self_double_colmajor) {
+TEST_F(TestRightMultiply_CSC_double, transpose_self_double_colmajor) {
     for (uint32_t key : {0}) {
         transpose_self(key, 30, 200, Layout::ColMajor, 0.01);
         transpose_self(key, 30, 200, Layout::ColMajor, 0.10);
@@ -313,7 +311,7 @@ TEST_F(TestRightMultiply_CSR_double, transpose_self_double_colmajor) {
     }
 }
 
-TEST_F(TestRightMultiply_CSR_double, transpose_self_double_rowmajor) {
+TEST_F(TestRightMultiply_CSC_double, transpose_self_double_rowmajor) {
     for (uint32_t key : {0}) {
         transpose_self(key, 30, 200, Layout::RowMajor, 0.01);
         transpose_self(key, 30, 200, Layout::RowMajor, 0.10);
@@ -321,7 +319,7 @@ TEST_F(TestRightMultiply_CSR_double, transpose_self_double_rowmajor) {
     }
 }
 
-TEST_F(TestRightMultiply_CSR_single, transpose_self_single) {
+TEST_F(TestRightMultiply_CSC_single, transpose_self_single) {
     for (uint32_t key : {0}) {
         transpose_self(key, 30, 200, Layout::ColMajor, 0.01);
         transpose_self(key, 30, 200, Layout::ColMajor, 0.10);
@@ -335,21 +333,21 @@ TEST_F(TestRightMultiply_CSR_single, transpose_self_single) {
 //
 ////////////////////////////////////////////////////////////////////////
 
-TEST_F(TestRightMultiply_CSR_double, submatrix_other_double_colmajor) {
+TEST_F(TestRightMultiply_CSC_double, submatrix_other_double_colmajor) {
     for (uint32_t key : {0}) {
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 0.1);
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 1.0);
     }
 }
 
-TEST_F(TestRightMultiply_CSR_double, submatrix_other_double_rowmajor) {
+TEST_F(TestRightMultiply_CSC_double, submatrix_other_double_rowmajor) {
     for (uint32_t key : {0}) {
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::RowMajor, 0.1);
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::RowMajor, 1.0);
     }
 }
 
-TEST_F(TestRightMultiply_CSR_single, submatrix_other_single) {
+TEST_F(TestRightMultiply_CSC_single, submatrix_other_single) {
     for (uint32_t key : {0}) {
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 0.1);
         submatrix_other(key, 3, 10, 5, 12, 8, 2, 1, Layout::ColMajor, 1.0);
@@ -363,106 +361,17 @@ TEST_F(TestRightMultiply_CSR_single, submatrix_other_single) {
 ////////////////////////////////////////////////////////////////////////
 
 
-TEST_F(TestRightMultiply_CSR_double, trans_other_times_sparse_colmajor) {
+TEST_F(TestRightMultiply_CSC_double, trans_other_times_sparse_colmajor) {
     uint32_t key = 0;
     transpose_other(key, 7, 22, 5, Layout::ColMajor, 0.05);
     transpose_other(key, 7, 22, 5, Layout::ColMajor, 0.10);
     transpose_other(key, 7, 22, 5, Layout::ColMajor, 0.80);
 }
 
-TEST_F(TestRightMultiply_CSR_double, trans_other_times_sparse_rowmajor) {
+TEST_F(TestRightMultiply_CSC_double, trans_other_times_sparse_rowmajor) {
     uint32_t key = 0;
     transpose_other(key, 7, 22, 5, Layout::RowMajor, 0.05);
     transpose_other(key, 7, 22, 5, Layout::RowMajor, 0.10);
     transpose_other(key, 7, 22, 5, Layout::RowMajor, 0.80);
-}
-
-////////////////////////////////////////////////////////////////////////
-//
-//     Public API: RandBLAS::spmm (dense * sparse)
-//
-//     Tests the public spmm overload: C = alpha * A * B + beta * C
-//     where A is dense and B is sparse.
-//
-////////////////////////////////////////////////////////////////////////
-
-template <typename T>
-class TestPublicAPI_DenseTimesSparse : public ::testing::Test {
-protected:
-    void SetUp() override {}
-    void TearDown() override {}
-
-    CSRMatrix<T> make_csr(int64_t m, int64_t n, T nonzero_prob, uint32_t key = 0) {
-        CSRMatrix<T> A(m, n);
-        std::vector<T> buf(m * n);
-        RandBLAS::RNGState s(key);
-        iid_sparsify_random_dense<T>(m, n, Layout::ColMajor, buf.data(), 1 - nonzero_prob, s);
-        dense_to_csr<T>(Layout::ColMajor, buf.data(), 0.0, A);
-        return A;
-    }
-
-    // Test: C = alpha * A * B + beta * C, where A is dense, B is sparse
-    void test_dense_times_sparse(
-        int64_t m, int64_t n, int64_t k,
-        T alpha, T beta,
-        Layout layout, T nonzero_prob, uint32_t key = 0
-    ) {
-        // Create dense A (m x k) and sparse B (k x n)
-        auto [A, A_layout, state] = random_matrix<T>(m, k, RandBLAS::RNGState(key));
-        auto B = make_csr(k, n, nonzero_prob, key + 1);
-
-        // Densify B for reference computation
-        std::vector<T> B_dense(k * n, 0.0);
-        to_explicit_buffer<T>(B, B_dense.data(), layout);
-
-        // Create initial C and copy for reference
-        auto [C_init, C_layout, state2] = random_matrix<T>(m, n, RandBLAS::RNGState(key + 2));
-        std::vector<T> C_ref(C_init);
-        std::vector<T> C_actual(C_init);
-
-        int64_t lda = (layout == Layout::ColMajor) ? m : k;
-        int64_t ldb = (layout == Layout::ColMajor) ? k : n;
-        int64_t ldc = (layout == Layout::ColMajor) ? m : n;
-
-        // Reference: blas::gemm with dense B
-        blas::gemm(layout, blas::Op::NoTrans, blas::Op::NoTrans,
-            m, n, k, alpha, A.data(), lda, B_dense.data(), ldb, beta, C_ref.data(), ldc);
-
-        // Actual: RandBLAS::spmm with sparse B
-        RandBLAS::spmm(layout, blas::Op::NoTrans, blas::Op::NoTrans,
-            m, n, k, alpha, A.data(), lda, B, beta, C_actual.data(), ldc);
-
-        // Compare results with tolerance scaled by k (inner dimension)
-        // Matrix multiply accumulates ~k operations, so error is O(k * eps)
-        T atol = k * std::numeric_limits<T>::epsilon();
-        T rtol = std::sqrt(std::numeric_limits<T>::epsilon());
-        test::comparison::buffs_approx_equal(
-            C_actual.data(), C_ref.data(), m * n,
-            __PRETTY_FUNCTION__, __FILE__, __LINE__,
-            atol, rtol
-        );
-    }
-};
-
-class TestPublicAPI_DenseTimesSparse_double : public TestPublicAPI_DenseTimesSparse<double> {};
-
-TEST_F(TestPublicAPI_DenseTimesSparse_double, basic_colmajor) {
-    test_dense_times_sparse(50, 30, 40, 1.0, 0.0, Layout::ColMajor, 0.10);
-    test_dense_times_sparse(50, 30, 40, 1.0, 0.0, Layout::ColMajor, 0.50);
-}
-
-TEST_F(TestPublicAPI_DenseTimesSparse_double, basic_rowmajor) {
-    test_dense_times_sparse(50, 30, 40, 1.0, 0.0, Layout::RowMajor, 0.10);
-    test_dense_times_sparse(50, 30, 40, 1.0, 0.0, Layout::RowMajor, 0.50);
-}
-
-TEST_F(TestPublicAPI_DenseTimesSparse_double, nontrivial_alpha_beta_colmajor) {
-    test_dense_times_sparse(50, 30, 40, 2.5, -1.0, Layout::ColMajor, 0.10);
-    test_dense_times_sparse(50, 30, 40, 2.5, -1.0, Layout::ColMajor, 0.80);
-}
-
-TEST_F(TestPublicAPI_DenseTimesSparse_double, nontrivial_alpha_beta_rowmajor) {
-    test_dense_times_sparse(50, 30, 40, 2.5, -1.0, Layout::RowMajor, 0.10);
-    test_dense_times_sparse(50, 30, 40, 2.5, -1.0, Layout::RowMajor, 0.80);
 }
